Refactor tests: Split them into one file per class, move them to test directory
